requirements.txt
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
et-xmlfile==2.0.0
fdb==2.0.4
firebird-base==2.0.2
firebird-driver==2.0.2
fonttools==4.62.1
gitdb==4.0.12
gitpython==3.1.46
greenlet==3.3.2
idna==3.11
jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
markupsafe==3.0.3
matplotlib==3.10.8
narwhals==2.18.0
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pillow==12.1.1
protobuf==5.29.6
pyarrow==23.0.1
pydeck==0.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
referencing==0.37.0
reportlab==4.4.10
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sqlalchemy==2.0.48
sqlalchemy-firebird==2.1
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
typing-extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
xlsxwriter==3.2.9
